modified user delete cont

+++ backend/controllers/admin/userDeleteController.js
@@ -4,17 +4,21 @@ const mongoose = require('mongoose');
 
 const deleteUser = async (req, res) => {
 	const { email, username } = req.body;
-	const param = email || username;
+	const query = email ? { email } : { username };
+	const value = email || username;
 	
 	try {
-		const foundUser = await User.findOne({ param });
+		const foundUser = await User.findOne(query);
 		const foundUserUid = foundUser.uid;
 		const deleted = await admin.auth().deleteUser(foundUserUid);
-		const dbDelete = await User.findOneAndDelete({ foundUserUid });
+		const dbDelete = await User.findOneAndDelete({ uid: foundUserUid });
 		
-		res.status(200).json({
-			message: `user ${param} deleted successfully`
-		})
+		if (deleted && dbDelete){
+				res.status(200).json({
+				message: `user ${value} deleted successfully`
+			});
+		}
+		
 	} catch (err) {
 		console.log(err);
 		res.status(500).send('Internal server error');
